completed cart update and cart load system. Runs but some things need to be fixed

+++ app/src/main/java/com/example/androideatit/Database/CartDataSource.java
@@ -17,7 +17,7 @@ public interface CartDataSource {
 
     Single<Integer> countItemCart(String uid);
 
-    Single<Long> sumPriceInCart(String uid);
+    Single<Double> sumPriceInCart(String uid);
 
     Single<CartItem> getItemInCart(String foodId, String uid);
